move rag_logging.py from Experiment folder to Backend/rag to make it in the application context
